requirements.txt
adjustText==0.7.3
anndata==0.6.22.post1
annoy==1.15.1
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bbknn==1.3.5
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cmake==3.13.3
colorlover==0.3.0
configobj==5.0.6
cryptography==2.7
cycler==0.10.0
Cython==0.29.10
decorator==4.4.0
defusedxml==0.5.0
DensityPlot==0.1.7
entrypoints==0.3
fbpca==1.0
fcsparser==0.2.0
FlowCal==1.2.1
future==0.16.0
globus-cli==1.10.1
globus-sdk==1.7.1
graphtools==0.1.9
greenlet==0.4.15
h5py==2.9.0
idna==2.8
imageio==2.6.1
importlib-metadata==0.18
intervaltree==2.1.0
ipyevents==0.6.3
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.2
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
jupytext==1.0.4
kiwisolver==1.0.1
leidenalg==0.7.0
lesscpy==0.13.0
llvmlite==0.28.0
loompy==2.0.17
louvain==0.6.1
magic==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
matplotlib-venn==0.11.5
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mnnpy==0.1.9.5
mock==2.0.0
more-itertools==7.0.0
mpld3==0.3
msgpack==0.6.1
MulticoreTSNE==0.1
natsort==6.0.0
nbconvert==5.4.1
nbformat==4.4.0
neovim==0.3.1
networkx==2.2
notebook==5.7.7
numba==0.43.1rc1
numexpr==2.6.9
numpy==1.15.4
packaging==19.0
palantir==0.2.1
palettable==3.3.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
patsy==0.5.1
pbr==5.1.3
pexpect==4.6.0
PhenoGraph==1.5.2
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.0.0
pluggy==0.12.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pycairo==1.18.1
pycparser==2.19
Pygments==2.3.1
PyGSP==0.5.1
PyJWT==1.7.1
pynvim==0.3.2
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.7.0
pytest==4.6.3
python-dateutil==2.8.0
python-igraph==0.7.1.post6
pytz==2018.9
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.22.0
retrying==1.3.3
rpy2==3.0.4
sam-algorithm==0.6.7
scanorama==1.0
scanpy==1.4.4.post1
scikit-image==0.16.2
scikit-learn==0.21.2
scipy==1.3.1
scvelo==0.1.16
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
statsmodels==0.10.1
style==1.1.0
tables==3.5.1
terminado==0.8.2
testfixtures==6.6.1
testpath==0.4.2
tornado==6.0.1
tqdm==4.32.2
traitlets==4.3.2
tzlocal==2.0.0b2
umap-learn==0.3.7
update==0.0.1
UpSetPlot==0.3.0.post3
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.2.2
zipp==0.5.1
